Implement DICompileUnit and DIFile
In order to generate debug information through LLVM, it is necessary to
declare a DICompileUnit in which all functions will live. Support for
doing this has now been added, though the file will always appear to be
"<stdin>" as it is not yet possible to assign a compile unit to a
specific source file.