Stub implementation of Cohen's display for supertypes
This is the initial stub of Cohen's display for tracking super types.
It only handles the Class supertypes so far and doesn't address
array classes or interfaces (yet).

It calculates some data on the sizes of the displays which will hopefully
come in handy in the future when compiling larger programs.

TODOs:
* interfaces
* array types
* investigate typeids for each class
* add a "depth" to each class
* update to use the new Logging support

Signed-off-by: Dan Heidinga <heidinga@redhat.com>