Remove LLBasicBlock implementation from FunctionDefinition
Previously, the FunctionDefinition interface extended the LLBasicBlock
interface. This was requiring needless extra checks elsewhere in the
code as to which implementation of LLBasicBlock was really being used
and introduced a lot of unnecessary pass-through methods in
FunctionDefinitionImpl.

Instead, FunctionDefinition now has a method to get the root block of
the function, which should be used instead.