Add an LLBuilder interface
Previously, construction of LLVM instructions was handled by methods on
LLBasicBlock. However, this required the caller to manually append debug
location information using meta("dbg", ...). This was both ugly and
error-prone.

LLVM's C and C++ interfaces instead use a dedicated builder object to
construct their instructions. This builder object can have a debug
location applied to it, which it will automatically apply to every
instruction constructed through it. This allows the debug location to be
set only once, which is much cleaner.

A new LLBuilder interface has been added to QCC to match this. This
allows LLVMNodeVisitor to set the debug location prior to visiting a
Node, meaning that the vast majority of code in that class no longer has
to worry about debug locations.

Closes: #194