Fix bad metadata on ret void and unreachable instructions
Previously, the ret void and unreachable LLVM instruction types used a
singleton instance when created. While this was fine most of the time,
it fell apart as soon as metadata was attached to one, as it would be
attached to the singleton instance and would thus actually be attached
to *all* instructions of that type. This has been corrected by removing
the singleton instances of these instructions and instead using new
instances each time.