Early access to VMHelpers class
This class is intended to be a home for methods that are required
by the Runtime to support the compiled code.

Currently it has the `fast` & `full` versions of instanceof but they
aren't yet functional and the signatures may change further in the
future.

Pushing now so other methods (ie: monitor support) can be homed here
as well if necessary.

Using this class requires having
`$HOME/.m2/repository/cc/quarkus/qcc-runtime-main/1.0.0-SNAPSHOT/qcc-runtime-main-1.0.0-SNAPSHOT.jar`
on the --boot-modules path as well

Signed-off-by: Dan Heidinga <heidinga@redhat.com>