Implement LLVM debug info for inlined call sites
In order to properly represent that inlining has occurred in debug
information, it is necessary to make use of the 'inlinedAt' attribute on
DILocation objects. Support for populating this attribute and for
ensuring correct filling of the 'scope' attribute for such DILocations
has now been added to the LLVM code generator.