Move all mapping logic out of LLVMGenerator
Previously, the LLVMGenerator class contained a lot of code for
performing mapping of certain globally visible values, like literals and
types. However, LLVMGenerator is not a good place for this as it is not
tied to a particular LLVM module. The introduction of identified types
to handle recursive type definitions will require that use of struct
types be limited to the module they were created in, which makes it
necessary to move this logic into a different class.

A new visitor class has been added to encapsulate the previous logic
that lives across an entire LLVM module rather than a single function.