Process line number tables at method resolution time
Previously, line number tables would only be processed and available for
searching during ilgen. This presented a problem for the LLVM codegen,
as it is necessary to give the DISubprogram representing each function a
line number to get correct line number information for instructions in
the function prologue.

To support this, logic for handling line number tables has been moved
into its own class and ExecutableElement will now keep track of the
minimum and maximum line numbers of the method.