Set proper line numbers for DISubprograms
Previously, all DISubprograms created by the LLVM codegen would be given
a line number of 0. This was done as there is no canonical way in Java
class files to represent what line a method declaration appeared on.
However, it turns out that LLVM will attempt to use this line number for
instructions that it generates in the function prologue. Furthermore,
LLVM uses a line number of 0 as a sentinel to suppress debug location
generation for an instruction. This was leading to instructions in the
function prologue getting the location data of the previous instructions
in memory, which was clearly incorrect.

Instead, the LLVM codegen will now use the minimum line number present
in the method (or line 1 if line number information is not available) as
the line number of the DISubprogram.

Fixes: #241