Split up different types of declarations in LLVM modules
Previously, different kinds of declarations in LLVM modules would be
emitted in whatever order they happened to be constructed. This would
mean mixing metadata, function declarations, and type declarations in a
difficult-to-read manner. This also caused an issue with a hidden
requirement in LLVM's parsing of LL files which requires that types be
fully declared before being used in a GEP, load, store, or alloca
instruction so that it can check that the type is sized.

This has been fixed by splitting LLVM module declarations into different
sections based on the type of declaration, which closely mirrors how
LLVM itself will emit LL files.

Fixes: #259