Assign `typeid` to the ValidatedTypeDefinition
This is a first pass at setting the typeid based on the work
done in the SupersDisplayTables.java plugin.  It only sets
the typeid and doesn't assign the bound but should be sufficent
for anyone that needs to use the typeids in other work.

A better solution that encapsulates more of the data - typeid,
range, and interface bits - will follow

Signed-off-by: Dan Heidinga <heidinga@redhat.com>