Add options to control generation of graph
Added following set of options to control generation of node graph of
the methods:
1. -g or --gen-graph : this is mandatory option if the user wants to
generate graph.
2. -m or --methods : this option allows to control the methods for
which graphs are generated. If not specified, then the default value is
all methods.
Format of the methods is:
	<package name>/<class name>.<method name><method descriptor>
        eg: my/example/package/Main.foo()V
3 -p or --phase : this option allows to control the compilation phase
for which graphs are generated. If not specified, then the default value
is all phases.

Examples:
- Generate graphs for all methods in all phases
	$ jbang cc.quarkus:qcc-main:1.0.0-SNAPSHOT -g <other args>

- Generate graphs for hello/world/Main.main()I method for all phases
	$ jbang cc.quarkus:qcc-main:1.0.0-SNAPSHOT -g -m
"hello/world/Main.main()I" <other args>

- Generate graphs for all methods in add and analyze phase
	$ jbang cc.quarkus:qcc-main:1.0.0-SNAPSHOT -g -p add,analyze
<other args>

Signed-off-by: Ashutosh Mehra <asmehra@redhat.com>