Track the `maximumSubtypeId` on VTDs
Update the VTD class and related classes to also support
setting the maximumSubtypeId.  This tracks the data on
the key class that's available to all plugins.

It also allows the lowering code to use it from there and for
all instanceof checks to have access to it.

Also add a `VTD[] getInterfaces()` method rather than having
to query them 1-by-1 based on the index.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>