Add options for enabling/disabling PIEs
Previously, QCC would not specify to the toolchain whether or not it
wanted to use PIEs. This could result in the linker and LLVM compiler
disagreeing on whether PIE support was required and causing linking to
fail. To correct this, new flags have been added to explicitly enable
and disable PIE support and both the LLVM compiler and the linker will
now explicitly have PIE support either enabled or disabled.

Fixes: #321