Implement collected pointer types
Previously, QBICC had two ways of representing pointers. One was with an
opaque reference type and the other was with a raw pointer. Operations
on reference types would need to be lowered to operations on raw
pointers, but this caused issues since it threw out the fact that the
raw pointer was a reference to a heap-allocated object, meaning that the
backend would have no way of distinguishing pointers that need to be
included in the GC maps from those that do not.

To bridge this gap, pointers can now be marked as collected, which will
communicate to the backend that they refer to objects on the heap and
thus need to be included in GC maps.