Modify ValueHandle to track pointer types instead of value types
Previously, ValueHandles would keep track of the type of the value they
referred to. However, this is no longer enough as it is necessary to
keep track of whether a pointer derived from a ValueHandle would be
collected or uncollected. To allow this, ValueHandle must keep track of
the type of pointer it would produce if used in an AddressOf node rather
than the type of value it would yield if used in a Load node.