Use addrspace(1) pointers for collected pointers and references
Previously, the LLVM codegen would use raw pointers for collected
pointers and would use plain integers for references. In order to avoid
inttoptr and ptrtoint casts (which often interfere with LLVM's alias
analysis) and just generally keep things clean, it would be better to
use pointers for references as well.

Additionally, since all pointer types in the IL are now known to be
either collected or uncollected, it is now possible to move to using
pointer types compatible with Statepoint's abstract machine model. This
is done by using address space 1 for all pointers that need to be
considered in the GC maps, which includes both collected pointers and
references. This does not change the semantics expected by the rest of
QBICC in any way and is generally only visible within the LLVM backend
itself.