Hide casts between pointers and references in pseudo-intrinsics
Previously, the addrspacecast instruction would be used to cast between
raw pointers and references/collected pointers. This was problematic as
doing this violates the Statepoint abstract machine's assumptions in
numerous ways. This could lead to a number of problems, including
invalid stack maps being generated and casts from references to raw
pointers being CSE'd across GC points.

This has been corrected by placing these addrspacecasts inside of opaque
functions that are not visible to LLVM prior to RS4GC running, called
pseudointrinsics. Currently, these pseudointrinsics are included in the
same LLVM module as the rest of the code. This is good enough for the
time being, but could cause problems if function attributes are inferred
or if the inliner is run. This could be corrected in the future by
placing them into a separate LLVM module file which is only linked into
the main one (via llvm-link) immediately before or after RS4GC.