Fix redundant bitcasts from i8* to i8*
The pseudointrinsic for converting a collected pointer into a raw
pointer always returns an i8*, which must be bitcast into the
appropriate type. There was a check that was meant to skip this if the
desired result type was already i8*, but this was not correctly being
triggered since some LLVM backend classes did not reimplement
Object.equals and would thus use object identity equality instead.