Clean up function/parameter attributes
Previously, there was some patchwork support for certain function and
parameter attributes. This was done differently in each location,
usually using enum values or a boolean flag, and was not deisnged to be
easily extended to support new attributes.

This has been replaced with a more robust system that allows arbitrary
LLValues to be used as attributes, which should allow more flexibility.