Always add the uwtable attribute
Previously, the uwtable attribute would only be added to functions
containing an invoke instruction. However, this is redundant (as the
unwind tables will be generated automatically in this case) and does not
generate unwind tables for functions without invokes in them. This is a
problem as the GC will need to use the unwind tables to walk the stack,
so any function that could potentially need to be walked during a stack
walk needs the uwtable attribute.