Deduplicate equivalent DILocations
Previously, the LLVM backend would always generate a new DILocation
whenever it was emitting code for a Node. This would result in a lot of
duplicate DILocations being emitted, as it was common for many Nodes to
use the same line number. To avoid this, non-distinct DILocations are
now deduplicated in LLVMModuleDebugInfo.