Add release fences after allocations
Previously, the nogc plugin would not perform any synchronization after
performing an allocation. This is a problem on weakly ordered
architectures, since the newly allocated object may immediately escape
to another thread running on another core without any synchronization.
In this case, the second thread may see uninitialized memory if the
escaping store occurs before the stores that initialize the object.

To combat this, a store-store fence is needed to ensure that the object
cannot be seen in an uninitialized state when escaping. This is done by
adding a release fence after each allocation.