Use NONE atomicity mode for initializing new objects
Previously, the nogc plugin used the UNORDERED atomicity mode for the
stores used to initialize objects. However, this is actually not
necessary since the object cannot have escaped the current thread to be
visible to any other threads by this point. Therefore, the NONE
atomicity mode will now be used for these stores.