RTA: improve precision of instance method reachability
An instance method of a class is not truly invokable until an
instantiation of that class is reached in an invokable method.  Add
the logic for deferring processing of such reachable, but not
instantiated instance methods.

Also make the RTA thread safe by synchronizing the operations
that make non-local changes to the class hierarchy.

Misc downstream fixes and adjustments to RTA clients.