Implement CProbe for functions
This patch extends CProbe to allow it to get the actual function name
if the function exposed as an API is just a C macro.
This is the case with many of libunwind APIs.
For example `unw_init_local` is defined as a macro which expands to a
platform specific function.
To get the actual function name CProbe would emit a file like:

	void * fn_address0 = <function_to_probe>;

Object file for this code would contain entry in symbol table for
`fn_address0` and its offset in data section.
For the same offset there would be an entry in the relocation section
and the symbol corresponding to that entry would give us the actual
function name.

Signed-off-by: Ashutosh Mehra <asmehra@redhat.com>