Add is_initialized(typeid) and set_initialized(typeid) intrinsics
We only need to store the fully initialized state in the native info as
that is enough for the fast-path success case to find it quickly.

For failed initialization, we want the call to happen in the state machine
and for the appropriate exception to be stored there, and thrown by it.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>