Add llvm-link stage
This patch adds a new stage in the compilation process.
The new stage uses llvm-link to combine all the bitcode files into
a single bitcode file.
This bitcode file is then compiled into assembly, then to an object file
and finally linked to get the executable.
The single bitcode file is named as "qbicc-app.bc" and is created in the
output directory specified by the user.

This step is required so that instruction offset for the stackmap
records are in increasing order, which would allow us to correctly
identify the callsite instruction.

Signed-off-by: Ashutosh Mehra <asmehra@redhat.com>