Implement @align and use to force maximal alignment on Unwind_Exception
Fixes crash on BigSur when throwing exceptions (eg TryCatch snippet).

The root cause of the problem was that the definition of _Unwind_Exception
in Apple's unwind.h was only forcing an 8 byte alignment of the struct,
while the specification (and the machinecode) of libunwind is expecting
a 16 byte alignment. The fix emulates the behavior of gcc's struct
definition which forces alignment to match the maximal platform alignment.
This isn't exactly what the IA64 C++ spec for libunwind says
(it requires "double word alignment"), but should be fine in practice.