Bug fix in conversion of invokes to linktime errors
An invokevirtual whose target is a native method cannot be converted
to an unconditional throw of an UnsatisfiedLinkError because the
virtual method may be overriden. Tested against the class library
at 8311cc4b2 (before the addition of a non-native Object.toString())
and verified that with this fix I see the same 4MB growth in final
executable size that I saw when defining a non-native Object.toString.