Improve reachability analysis of interpreter heap
Replace the conservative assumption that all types that were
instantiated in the interpreter's heap are reachable with the
proper optimistic algorithm that instead traces the portions of
the heap that can be reached from static fields as those static
fields become reachable. We will inspect each object in the
interpreter heap at most once during this analysis (once an
object is reachable from one static, it doesn't matter if
it is also reachable from other statics).