One more fix for small type IDs
It papers over another problem which is that an implicit cast from (small) type_id to short integer results in a `bitcast` of `i16` to `i16` at the LLVM level, so that would need investigating at some point.  But, this solution slightly improves the existing code anyway so we can solve the other later.