fix invokability definition for superclass methods
Because we are building "classic" class-based vtables, if an
instance method of a class that overrides a superclass method
is invokable, we have to consider the superclass method invokable
(even if it is never invoked at the superclass static type). This
ensures a compatible vtable layout for all related classes/methods.