Reintroduce resolver pattern for parameters
This is a step towards cleaning up the signature/type annotation mess.  The existing mechanism requires that parameter elements be constructed before their enclosing methods, which creates a problem where the enclosing element cannot be acquired from the parameter, which in turn means that the enclosing method cannot be used to resolve type parameters.  `TypeParameterContext` was an effort to solve the problem but it did not really work out.  By resolving parameters late, we can add the enclosing element and also potentially drop `TypeParameterContext` in the future.