Propagate interface/abstract method escape analysis
* ParameterValue now takes element into account
when computing equals. This means that for
interface or abstract/overriden method calls,
multiple instances exist for the same parameter
index slot.
* If any of them escapes globally,
that information needs to be propagated to other
ParameterValue instances for escape analysis to
work as expected.
* Method exit is a good place to add this,
because it already takes care of propagating
global escape information.
* The fix extends this propagation to other
ParameterValue instances.