Switch to more compact graph output
* Each node represents now a basic block.
* Edges represent basic block connections,
coming out of terminators and support EA info
(e.g. points-to and fields edges).
* Added EA information to new graphs.
* Add block info index to basic block,
and consume that from the disassembler.
* Some nodes are inlined into expressions,
while some others are assigned an identifier
with the format `b<block_index>.<node_id>`.
The decision is a bit subjective,
but readability of graph has been the key driver.
* Identifiers are used instead of edges
in order to keep the graph as readable as possible.