move StaticFieldElements->GlobalVariables map from Lowering to BuildtimeHeap
The static fields of the program (along with the Class objects, and the
ObjectLiterals embedded in the program text) are the roots of the BuildtimeHeap.
Make the BuildtimeHeap responsible for managing them. This is motivated
by the need to be able to lower StaticFieldPointers that are contained in
instance fields of the serialized BuildtimeHeap objects without introducing
a circular dependency between plugins.