add intrinsic to get a C function pointer to a Java static method
Use case: passing a pointer to an @export function as a callback
down to a CNative function. By having a generic intrinsic
for this, we can be more flexible in where these functions are defined.

I explored trying to do addr_of_function on a Class::methodr instead of
adding an intrinsic, but the invokedynamic machinery causes enough
indirection that it seemed overly challenging to identify the pattern
and ensure that we get the function pointer we need (one that goes
directly to the function with the @export annotation).