restructure feature processing to avoid premature loading
Ensure that all of the "patches" implied by GraalVM and qbicc features
are processed before doing any of the qbicc classloading of the
application classes that depends on those patches. This makes the
patches actually take effect (installing a patch after the DTD is
loaded to make a LTD is ineffective).