hooks for resource loading
Hook the three key methods on ClassLoader for loading resources.
We can trivial implement getResourceAsStream, so do that now.
The remaining two methods findResource and findResources require real
VFS integration, so in this commit we simply print out a warning
message when there is a findable resource that we are ignoring.
Also add a simple integration test for getResourceAsStream.