Restructure reflection registration/generation
Changes to support registering core JDK classes for reflective access
via a qbicc Feature. This is needed to support serialization in
Jackson (and probably will be generally useful for other frameworks).

The key change is to ensure that a qbicc feature can be processed
without requiring any types to be loaded.  Processing the feature only
populates a registry with String-based lookup structures that are then
queried as types are loaded.  It is only on actual loading of a type
that we handle the reachability implications of these annotations and
spawn off tasks to generate the required Method/Field/Annotation/etc
instances in the build time heap.